Update Java version and create Pong:jar artifact
Adjusted the project's Java version in the ProjectRootManager from liberica-1.8 (2) to liberica-1.8 for consistency. Also, a new artifact named 'Pong:jar' has been added to manage the output. An unneeded import statement was removed from Server.java and a new MANIFEST.MF file was created in the META-INF directory to specify GameGUI.Game as the main class.